// CabBookingSPA/src/app/shared/models/bookingDetailsModel.ts
// Generated by https://quicktype.io

export interface BookingDetails {
    id:              number;
    email:           string;
    bookingTime:     string;
    pickupTime:      string;
    contactNo:       string;
    status:          string;
    locationRequest: LocationRequest;
}

export interface LocationRequest {
    id:                  number;
    pickupAddress:       string;
    dropoffAddress:      string;
    landmark:            string;
    placeIdRequestModel: PlaceIDRequestModel;
}

export interface PlaceIDRequestModel {
    id: number;
}